UAK : add support for access key checking
This commit includes the implementation of methods that will allow
for access key checking during activation. The key is retrieved
from the VPD, and then checked against the BUILD_ID. If the
necessary criteria is met, the activation process continues as
requested. If not, the activation process is aborted and an error
is displayed.

Change-Id: I7d76f76743bdd73dc8446d71c79fb265b2c4811c
Signed-off-by: Zami Seck <zimzam17@gmail.com>