uak: Add sync method
Upon start, the software updater service reads the User Access Key
date from the com.ibm.ipzvpd.UTIL D8 property on the bmc card and from
flash memory in /var/lib/phosphor-bmc-code-mgmt/uak-data.
Then it syncs them up with one another, prioritizing first the bmc card,
then flash memory.

Tested: Logged in to the phyp console and updated the UAK date. Verified
that this registered in the D8 property immediately and that it was
written to flash memory upon reboot. Verified that syncing happens in
order of priority. Verified no syncing is done when the value is null,
0s, or spaces.

Change-Id: I6052f370d2168c45cf4db8d7809abeaa8035d650
Signed-off-by: Isaac Kurth <blisaac91@gmail.com>
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>