UAK Validation: Change error thrown
This commit changes the error reported when the update access key
requirements are not met. Instead of reporting an "InternalError"
the newly defined ExpiredAccessKey error is thrown in its place.
This error not only allows there to be a more specific error, there
is also metadata added to provide more information to the user.
This commit also changes the error reported during signature
verification from InternalError to InvalidSignature.

Change-Id: Id3ac412a3f0d3d3972d9701d8cd28bbaba9a7b6c
Signed-off-by: Zami Seck <zimzam17@gmail.com>