Initiate Dump After Code Update Failure (#27)
This change ensures that if code update fails due to one of the
checks (MSL, UAK verification, Signature Validation), a dump
is initiated. Because the errors that are triggered during these
checks changed from internal failure to more specific errors,
whenever a failed code update occurred, there was no longer a dump.
Now there is a dbus call to create a dump whenever necessary.

Co-authored-by: zamiseck <zimzam17@gmail.com>