Add Progress Code For USB Code Update (#31)
This change ensures that update progress codes are shown on
the op panel during the process of a firmware change. There
are four update codes which indicate that the firmware is:
1. C100B101: The firmware image is being installed on a side of
the flash.
2. C100B102: The firmware image is being installed on the other
side of the flash.
3. C100B103: The firmware installation has been completed
successfully.
4. C100B104: code is displayed, the firmware installation has
failed.
This reassures the user that the code update process is happening
and they will not need access to the GUI for further information.

Tested:
Made sure that the appropriate code was displayed on the op panel
and log within the progress code logs during code update.

USB Code Update: Cleanup (#39)

* USB Code Update Cleanup: Fail Code

These changes ensure that the fail code that should be displayed
when a USB code update attempt is only displayed then. Previously,
the C100B104 fail could be seen on the op panel during other types
of updates as well as USB code update. These changes ensure that
it is only visible when appropriate.

* USB Code Update Cleanup: Clearing Directory

When USB code update fails and is later retried, the update begins
updating the second side rather than starting the code update
from the beginning. These changes address that issue by clearing
directories after failure so upon reattempt, the USB code update works
as expected.

* USB Code Update Cleanup

Because the fail progress code is no longer being called from the
activation.cpp file, there is no longer a need for the progress
code logic to be in utils. These changes move the progress code calls
from the utils files to the USB manager files.

Change-Id: I46572f061b6a9d02ad2acea97cd212b61dcdd3b4
Signed-off-by: zamiseck <zimzam17@gmail.com>