Add Progress Code For USB Code Update (#31)
This change ensures that update progress codes are shown on
the op panel during the process of a firmware change. There
are four update codes which indicate that the firmware is:
1. C100B101: The firmware image is being installed on a side of
the flash.
2. C100B102: The firmware image is being installed on the other
side of the flash.
3. C100B103: The firmware installation has been completed
successfully.
4. C100B104: code is displayed, the firmware installation has
failed.
This reassures the user that the code update process is happening
and they will not need access to the GUI for further information.

Tested:
Made sure that the appropriate code was displayed on the op panel
and log within the progress code logs during code update.

Co-authored-by: zamiseck <zimzam17@gmail.com>