Ensure /tmp/images/ Directory Does Not Fill Up (#32)
This change ensures that when multiple images are uploaded to
the /tmp/images/ directory for code update--causing the device to
run out of space and an untar error--the /tmp/images/directory is
cleared, allowing the user to reupload the image without needing
to reboot.

TESTED:
uploaded multiple images until the bmc filled with space and the
untar error is displayed in the journal. Checked to see that the
/tmp/images/ directory still existed but was empty. Then, with
the cleared directory, uploaded another image and ensured that
the code update good path still works as expected.

Change-Id: Ib769aefb93ad1e0aac3a4a6faa3ace4c37aab3cf
Signed-off-by: zamiseck <zimzam17@gmail.com>