Ensure only one code update occurs at a time (#33)
This change ensures that when a code update is in progress,
another image can not be updated as well. This is important
because when two code updates are started one after the other it
causes failures, unintended write overs and leaves the machine in
a bad state.

TESTED:
Flashed two images onto the bmc. Started the code updates one after
the other. Ensured that the second update fails and the busy failure
error is seen in the journal.

SW551298

Co-authored-by: zamiseck <zimzam17@gmail.com>