Ensure only one code update occurs at a time (#33)
This change ensures that when a code update is in progress,
another image can not be updated as well. This is important
because when two code updates are started one after the other it
causes failures, unintended write overs and leaves the machine in
a bad state.

TESTED:
Flashed two images onto the bmc. Started the code updates one after
the other. Ensured that the second update fails and the busy failure
error is seen in the journal.

Change-Id: Iffb1c45ba20005b1b87303fdc2299f67dc4f9a38
Signed-off-by: zamiseck <zimzam17@gmail.com>

activation: Remove the BusyFailure error log (#6)

Just log a trace, since creating errors from a local yaml file and from
phosphor-dbus-interfaces in the same file is not supported.
A new Busy failure would need to be added to phoshpor-dbus-interfaces to
support logging an error.

Change-Id: I0c8e14c8ad5c5a37fbb62e3c2d2f0e6ecc4cd8ee
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>