UAK Emergency Bypass Fix (#40)
This change ensures that if the image being uploaded is an
emergency service pack, it will activate despite the
presence of the update access key date or not. This is necessary
because there may be certain instances where the machine has
lost that information. If that is the case, code update would be
impossible if the machine is in fieldmode.

Change-Id: Iafd551adc19a53a6a0ab56dcadba89959f583661
Signed-off-by: Zami Seck <zimzam17@gmail.com>