Clear /tmp/images/ Directory Upon Reboot
This change ensures that upon reboot. The /var/tmp/images/ and
the symlinked /tmp/images/ directory is cleared. When /tmp/images/
did rely on symlinks for storage, it was automatically cleared upon
reboot and code update. With the symlink change, it was necessary to
add extra logic to ensure the same.

TESTED:

- Ensured scp code update worked as expected despite the symlink to
/tmp/images/ from /var/tmp/images/

- Ensured redfish code update worked as expected despite the sympink to
/tmp/images/ from /var/tmp/images

- Ensured directories were cleared upon reboot and code update.

Change-Id: Id556d4d2af893252a66e9b64e3c4a0e6ba9e8f06
Signed-off-by: zamiseck <zimzam17@gmail.com>