ensure disable of reboot occurs after enable
Seen a few instances where both the enable and disable service seem to
get queued up at the same time (at least so close together that systemd
sees them as being requested to run at the same time). This results in a
race condition on which runs first. Bad things happen if the disable
happens to get called before the enable.

Put an explicit dependency in the services so if they are both queued at
the same time, they are run in the correct order.

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>
Change-Id: I0c7a04ede468e200b7fa2ca8184907ebeaf82d5d