Marker LID Validation : define validation dbus interface
This commit defines the dbus interface “validate” that will be
called during inband code updates. As of now, the interface checks
the validity of the access key. If the access key has expired and
field mode is enabled, the dbus returns an error specifying the
problem.

Tested:
Used busctl call “busctl call xyz.openbmc_project.Software.BMC.Updater
/xyz/openbmc_project/software xyz.openbmc_project.LID.Validate
Validate s /media/hostfw/running-ro/80a00001.lid” to ensure that the
call performs the validation call and the appropriate returns and
errors are logged.

Signed-off-by: Sagar Srinivas <sagar.srinivas@ibm.com>