msl: reset: Set MSL to version of subsequent update (#51)
Currently, a MSL reset will write the GA level to VPD to allow all
downgrades to be performed.
Change it so that the reset request creates a temporary file that would
signal the code update process to write the MSL to the version of the
subsequent update. Therefore when a reset MSL is issued, and a system is
downgraded to the first service pack for example, it would write that
level to VPD, preventing further downgrades to the GA level.

Tested: Verified the reset MSL allowed a downgrade and that the level
was written to VPD.
- First check the downgrade is blocked:
Apr 04 18:26:32 p10bmc phosphor-image-updater[547]: BMC Minimum Ship
Level (fw1040.00-20                    ) NOT met by
fw1020.30-24-1020.2314.20230329a (NL1020_099)
- Issue reset request:
root@p10bmc:~# software-manager-tool --resetminlevel
- Retry the update:
Apr 04 18:27:23 p10bmc phosphor-image-updater[547]: Resetting Minimum
Ship Level to: fw1020.30-24
Apr 04 18:27:31 p10bmc systemd[1]: Starting Write image 86c7bbd9 to
BMC storage...

Change-Id: Ic37d977cfad720205ea13c3894458d4c7abb104a
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>