Marker LID Validation : Define validation dbus interface Backend (#17)
This commit defines the dbus interface “validate” that will be
called during inband code updates. As of now, the interface checks
the validity of the access key. If the access key has expired and
field mode is enabled, the dbus returns an error specifying the
problem.

This commit also adds restriction on HIPER service packs, being
allowed across GA levels. HIPER service packs only within the same
GA level will be allowed to be updated.
Example: If a customer is on a build with the GA level of fw1030, they
cannot move to the HIPER service pack released for fw1050.

Previously, HIPER service packs could be applied across GA levels
and this allowed customers to utilize the features of a latter service
pack without buying the service contract.

Tested:
Used busctl call “busctl call xyz.openbmc_project.Software.BMC.Updater
/xyz/openbmc_project/software xyz.openbmc_project.LID.Validate
Validate s /media/hostfw/running-ro/80a00001.lid” to ensure that the
call performs the validation call and the appropriate returns and
errors are logged.

Signed-off-by: Sagar Srinivas <sagar.srinivas@ibm.com>