start-update-interface: add the new app
Add the meson option for the start update D-Bus interface
implementation. Add the SoftwareManager new daemon to run in parallel
with existing code. The intention for new daemon is to combine
ImageManager and ItemUpdater into one. After due testing and
verifications ImageManager and ItemUpdater would be retired for
SoftwareManager. For more more details on design refer to -
https://gerrit.openbmc.org/c/openbmc/phosphor-dbus-interfaces/+/65738
https://gerrit.openbmc.org/c/openbmc/docs/+/65739

Change-Id: Id11823ca0ff62e76595aa4ad33212a2a2428c261
Signed-off-by: Jagpal Singh Gill <paligill@gmail.com>