msl: Add sync method
Upon start, the software updater service reads the Minimum Ship Level
from the com.ibm.ipzvpd.VSYS FV property on the bmc card and from
flash memory in /var/lib/phosphor-bmc-code-mgmt/msl-data.
Then it syncs them up with one another, prioritizing first the bmc card,
then flash memory.

The sync mechanism is needed to keep a backup of the value in order to
restore it if the planar or BMC card are replaced.

Tested: Verified the syncing is done upon reboot. Verified no syncing is
done when the value is all spaces (default), and when the VPD service is
not running.

Change-Id: I996f373711ac958a98e01fd65b09203016cbd69b
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>