msl: Ignore write errors during sync (#47)
The IBM VPD service may not be available when the code update service
starts, so ignore the VPD write errors. Add a sync call during the
verify call that is made during a code update (the uak is also synced as
part of the code update flow), and a sync call when the msl value is set
by the user to ensure the bakcup copy is created when the value is set.

Tested: Verified there are no dumps created when the BMC boots up even
if the VPD service has not started.

Change-Id: I314788ba7f164a501909073cc7ef61b77abcfe00

Change-Id: I362c9412aeef88c4a6f5334c4119c69569b61f84
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>