lid: validate: Fix parsing logic
Parse all the marker lid fields first before doing the verifications
because the uak check relies on the SPNM field, which currently is
placed after the FIPP field, causing the UAK check to have an empty
version string.

If the version string was empty, the UAK check was not logging an error.
Need to log an error to tell PHYP that the update has failed. Although
an empty version string should not be seen again unless there's an error
in the marker lid.

For the msl, read version string before appending null terminator.
Need to add the null terminator after reading the string, otherwise the
read operation will overwrite the null terminator.

Change-Id: Icb84873836e1a44cafc489b45b39d90701e84e32
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>