mmc: Create a log when factory reset is requested (#55)
For debug, it has been requested to create an informational error log
(PEL) that calls home when a factory reset is requested. This log will
need to be offloaded by the HMC prior to the BMC reboot.

The HMC takes less than a second to detect a new log, offload it, and
ack it. Therefore sleep for 3s to allow enough time for this process to
happen. The processing thread in the HMC may be busy if there are other
events being created, but from testing with multiple logs created, the
HMC handles them still within a second.

It is not critical to ensure that this log is offloaded by the HMC on
all cases. Therefore it is not worth the effort on trying to determine
if the system is HMC-managed and track that the HMC has ack'd the log.

Change-Id: Id43b0d238193f095bf2666b54966d8ed2710bcb7

Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>